requirements.txt
python-dotenv
telebot
google-cloud-compute
google-protobuf